+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/networks/cargo/CargoNetworkTask.java
@@ -64,7 +64,7 @@ class CargoNetworkTask implements Runnable {
         long timestamp = System.nanoTime();
 
         // Chest Terminal Code
-        if (SlimefunPlugin.getThirdPartySupportService().isChestTerminalInstalled()) {
+        if (SlimefunPlugin.getIntegrations().isChestTerminalInstalled()) {
             network.handleItemRequests(inventories, chestTerminalInputs, chestTerminalOutputs);
         }
 
@@ -85,7 +85,7 @@ class CargoNetworkTask implements Runnable {
         }
 
         // Chest Terminal Code
-        if (SlimefunPlugin.getThirdPartySupportService().isChestTerminalInstalled()) {
+        if (SlimefunPlugin.getIntegrations().isChestTerminalInstalled()) {
             // This will deduct any CT timings and attribute them towards the actual terminal
             timestamp += network.updateTerminals(chestTerminalInputs);
         }
