+++ src/main/resources/plugin.yml
@@ -1,13 +1,17 @@
+# Name and version
 name: Slimefun
 version: ${project.version}
 
+# Project metadata
 author: The Slimefun 4 Community
 website: https://github.com/Slimefun
 description: Slimefun basically turns your entire Server into a FTB modpack without installing a single mod
 
+# Technical settings
 main: io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin
 api-version: '1.14'
 
+# (Soft) dependencies of Slimefun, we hook into these plugins.
 softdepend:
 - CS-CoreLib
 - PlaceholderAPI
@@ -16,12 +20,19 @@ softdepend:
 - mcMMO
 - ItemsAdder
 
+# We hook into these plugins too, but they depend on Slimefun.
+loadBefore:
+- ChestTerminal
+- ExoticGarden
+
+# Our commands
 commands:
   slimefun:
     aliases: sf
     description: Slimefun base command
     usage: You either forgot to install CS-CoreLib or you installed an unsupported version.
 
+# Default permissions
 permissions:
   slimefun.command.guide:
     description: Allows you to obtain the Slimefun guide book
